# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "URL Shortener Service",
        "title": "Shortener API",
        "contact": {
            "email": "kokoulin92@gmail.com"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "paths": {
        "/": {
            "post": {
                "description": "method to get a single long url by a short url",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "method to save a single url",
                "operationId": "createShortURL",
                "parameters": [
                    {
                        "description": "Contains a string with an url",
                        "name": "url_data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "307": {
                        "description": "Temporary Redirect",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "the parameter is missing",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "the parameter not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "410": {
                        "description": "the parameter was deleted",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/{id}": {
            "get": {
                "description": "method to get a single long url by a short url",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "method to get a single long url",
                "operationId": "retrieveShortURL",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ShortURL",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "307": {
                        "description": "Temporary Redirect",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "the parameter is missing",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "the parameter not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "410": {
                        "description": "the parameter was deleted",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    }
}